requirements.txt
transformers
streamlit
python_dotenv